# schema/openapi.yaml
openapi: "3.0.0"
info:
  version: "0.1"
  title: "arXiv Compliler Service"
  contact:
    name: "arXiv API Team"
    email: nextgen@arxiv.org
  license:
    name: MIT

components:
  parameters:
    source_id:
      name: source_id
      in: path
      description: The upload workspace ID.
      required: true
      schema:
        type: integer
    checksum:
      name: checksum
      in: path
      description: The checksum of the upload workspace being compiled.
      required: true
      schema:
        type: string
    format:
      name: format
      in: path
      description: The compilation output format.
      required: true
      schema:
        type: string

paths:
  /status:
    get:
      operationId: getServiceStatus
      summary: |
        Get information about the current status of compilation service.
      responses:
        '200':
          description: OK
        '503':
          description: Service Unavailable

  /{source_id}:
    parameters:
    - $ref: '#/components/parameters/source_id'
    get:
      operationId: getCompilationStatus
      description: Return information about all compilations requested for the upload workspace.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'resources/compilationInfo.json'


  /{source_id}/{checksum}/{format}:
    parameters:
    - $ref: '#/components/parameters/source_id'
    - $ref: '#/components/parameters/checksum'
    - $ref: '#/components/parameters/format'
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'resources/compilationProduct.json'
        '303':
          description: Redirect to getTaskStatus for the compilation.
        '404':
          description: No compilation has been requested.
    post:
      operationId: requestCompilation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'resources/requestCompilation.json'
      responses:
        '202':
          description: Accepted for compilation.
          headers:
            Location:
              description: The URL for the compilation task.
              schema:
                type: string

  /{source_id}/{checksum}/{format}/log:
    parameters:
    - $ref: '#/components/parameters/source_id'
    - $ref: '#/components/parameters/checksum'
    - $ref: '#/components/parameters/format'
    get:
      operationId: getCompilationLog
      responses:
        '200':
          description: OK

  /{source_id}/{checksum}/{format}/product:
    parameters:
    - $ref: '#/components/parameters/source_id'
    - $ref: '#/components/parameters/checksum'
    - $ref: '#/components/parameters/format'
    get:
      operationId: getCompilationProduct
      responses:
        '200':
          description: OK

  /task/{uuid}:
    parameters:
    - name: uuid
      in: path
      description: The task's UUID.
      required: true
      schema:
        type: string
    get:
      operationId: getTaskStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: 'resources/compilationProduct.json'
        '301':
          description: With the task complete, redirect to the completed object.
          headers:
            Location:
              description: The URL for the successfully completed object.
              schema:
                type: string